Fixed crashing when we try to cast Integer to Double
Again, primitives vs objects

+++ src/pl/cba/genszu/amcodetranslator/interpreter/variable/types/IntegerVariable.java
@@ -221,7 +221,7 @@ public class IntegerVariable extends Variable {
 	}
 
 	public double toDouble() {
-		return (double) this.getValue();
+		return (double) (Integer) this.getValue();
 	}
 
 	public Variable convert(String type) {
